update string test
