--- test/cases_source_type_string/1.1.btx_log.txt
@@ -1,1 +1,0 @@
-event_2: { sf_3 = "" }
