--- test/cases_source_type_string/1.2.btx_log.txt
@@ -1,1 +1,0 @@
-event_2: { sf_3 = "hyEhk9uaOQJb = 8HT+0u/lu6rGq4ivNLx8HSYaiWvVfRW6fkDQX16FgJtv7mIHRAQW0mLUnYySLEB0X/ptpitF+RuAfWv6hWy/pW/CFpdbiESVPxB7nm8zgeTWfORTPKTRZZ0yNrcHMDdDjTlaEHl+eh+0drHekW0GyAGyioKx7wEtdYYyHJP1sOjWQDxEeXcmvvBJRsRvpnbMYh GVtq5YJ   JLCZs/ENyt0e9bmN5jCbgKUauqRr78Jx8L6uonBFueYr2ge6ibjSWM=" }
