--- test/cases_source_type_string/1.btx_model.yaml
@@ -1,11 +1,0 @@
----
-:stream_classes:
-- :name: sc1
-  :event_classes:
-  - :name: event_2
-    :payload_field_class:
-      :type: structure
-      :members:
-      - :name: sf_3
-        :field_class:
-          :type: string
